feat: move Trie to datastructures subfolder
