+++ src/main/java/com/thealgorithms/datastructures/tries/Trie.java
@@ -1,46 +1,4 @@
-package com.thealgorithms.tries;
-
-import java.util.HashMap;
-import java.util.Map;
-
-/**
- * TrieNode class which holds the characters and references to its child nodes
- */
-class TrieNode {
-    private Map<Character, TrieNode> children;
-    private char letter;
-    private boolean end;
-
-    TrieNode(char letter) {
-        this.letter = letter;
-        this.children = new HashMap<>();
-        this.end = false;
-    }
-
-    public Map<Character, TrieNode> getChildren() {
-        return children;
-    }
-
-    public void setChildren(Map<Character, TrieNode> children) {
-        this.children = children;
-    }
-
-    public char getLetter() {
-        return letter;
-    }
-
-    public void setLetter(char letter) {
-        this.letter = letter;
-    }
-
-    public boolean isEnd() {
-        return end;
-    }
-
-    public void setEnd(boolean end) {
-        this.end = end;
-    }
-}
+package com.thealgorithms.datastructures.tries;
 
 /**
  * Trie class which holds Strings of characters.
